Fix string equals if other is null and string constant pool
The loop from internedString to host and back to guest crashes the constant handling in the constant pool.
I think, it should be sufficient to just hand over the internedString to SPouT and add it to the pool.
String equals has not gracefully handled comparison with null in all cases